retrive recommendation complete

+++ lib/movies/data/models/movie_detail_model.dart
@@ -22,5 +22,5 @@ class MovieDetailModel extends MovieDetail {
           releaseDate: json["release_date"],
           runtime: json["runtime"],
           title: json["title"],
-          voteAverage: json["vote_average"]);
+          voteAverage: json["vote_average"].toDouble());
 }
